add .env file

+++ main.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/developer-afo/go-fiber-postgres/models"
-	"github.com/developer-afo/go-fiber-postgres/storage"
+	"github.com/developer-afo/go-fibre-postgres/models"
+	"github.com/developer-afo/go-fibre-postgres/storage"
 
 	"github.com/gofiber/fiber/v2"
 	"github.com/joho/godotenv"
@@ -50,11 +50,19 @@ func (r *Repository) CreateBook(context *fiber.Ctx) error {
 func (r *Repository) DeleteBook(context *fiber.Ctx) error {
 	bookModel := models.Books{}
 	id := context.Params("id")
+
 	if id == "" {
 		context.Status(http.StatusInternalServerError).JSON(&fiber.Map{
 			"message": "id cannot be empty",
 		})
 		return nil
+	}
+
+	n_err := r.DB.Where("id = ?", id).First(bookModel).Error
+	if n_err != nil {
+		context.Status(http.StatusBadRequest).JSON(
+			&fiber.Map{"message": "could not get the book"})
+		return n_err
 	}
 
 	err := r.DB.Delete(bookModel, id)
@@ -66,7 +74,7 @@ func (r *Repository) DeleteBook(context *fiber.Ctx) error {
 		return err.Error
 	}
 	context.Status(http.StatusOK).JSON(&fiber.Map{
-		"message": "book delete successfully",
+		"message": "book deleted successfully",
 	})
 	return nil
 }
